Massive Re-org
This is step 1 to re-writting the tank game.
2 major changes:
 - moved all the files into folders that make sense
 - adding glad to bring in all new GL functions

The move of files is to get all the source out of the main
repo folder. There were starting to be a lot of files and
it was getting out of hand.

The addition of glad is for the various GL extensions that
are needed. It will load all the extensions that is asked
for and give the apporopriate functions to be used in code.

After this major commit, work will start on making the
same functionality but in a new code base that is organised
better. The old code will be copied over and used in the new
code as it is correct in the conception, but uses a dead end
methodology of GL and the structure of rendering reflects that.