Basic window test setup
Wrote a basic main.cpp to simply test the window class.
It displays nothing but does exit correctly.

CMake conversion for the build system.
This handles the build, and should be extensible.

Moved the header files into an include folder.
This is to match more standard organisation styles
and it matches nicely with the way CMake is setup.

Minor bug fixes in Window.

Signed-off-by: glop102 <alexlikeschess@gmail.com>